% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_GBIF.R
\name{Plot_GBIF}
\alias{Plot_GBIF}
\title{Find the species from GBIF present in a plot}
\usage{
Plot_GBIF(polygon, EPSG, file_name)
}
\arguments{
\item{polygon}{An sf object of a polygon.}

\item{EPSG}{A 4-digit code for the projected coordinate system related to the sf object in meters.}

\item{file_name}{A file name "" you want to save the final csv to.}
}
\value{
Species (if present) in a plot with their IUCN and residence status
}
\description{
The function takes a plot as input and searches for any species
from the Global Biodiversity Information Facility (GBIF) that
are either present within the plot or located nearby. It then
determines the associated International Union for Conservation
of Nature (IUCN) status of the identified species and indicates
whether they are permanent or temporary residents, if applicable.
}
