% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Natura2000.R
\name{Plot_Natura2000}
\alias{Plot_Natura2000}
\title{Find the species from Natura2000 present in a plot}
\usage{
Plot_Natura2000(polygon, Natura, species, file_name)
}
\arguments{
\item{polygon}{An sf object of a polygon.}

\item{Natura}{An sf object with Natura2000 sites.}

\item{species}{A csv file with the species associated to Natura2000 sites.}

\item{file_name}{A file name "" you want to save the final csv to.}
}
\value{
Species (if present) in a plot with their IUCN and residence status
}
\description{
The function takes a plot as input and searches for any species
from the Natura2000 sites that are either present within the plot
or located nearby. It then determines the associated International
Union for Conservation of Nature (IUCN) status of the identified
species.
}
