% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_GBIF.R
\name{Plot_GBIF}
\alias{Plot_GBIF}
\title{Extract GBIF species occurrence data from a polygon and get IUCN Red List status}
\usage{
Plot_GBIF(polygon, EPSG, file_name)
}
\arguments{
\item{polygon}{A spatial polygon object (class "sf") representing the study area}

\item{EPSG}{A 4-digit code for the projected coordinate system related to the sf object in meters}

\item{file_name}{A character string representing the name of the output file}
}
\value{
A data frame (if species are present) with three columns: scientific name, IUCN Red List status, and Residence
}
\description{
The function takes a plot as input and searches for any species
from the Global Biodiversity Information Facility (GBIF) that
are either present within the plot or located nearby. It then
determines the associated International Union for Conservation
of Nature (IUCN) status of the identified species and indicates
whether they are permanent or temporary residents, if applicable.
}
