% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_GBIF.R
\name{Plot_GBIF}
\alias{Plot_GBIF}
\title{Extract GBIF species occurrence data from a polygon and get IUCN Red List status}
\usage{
Plot_GBIF(polygon, EPSG, file_name)
}
\arguments{
\item{polygon}{A spatial polygon object (class "sf") representing the study area}

\item{file_name}{A character string representing the name of the output file}

\item{epsg}{A 4-digit code for the projected coordinate system related to the `polygon` object in meters}
}
\value{
A data frame with three columns: scientific name, IUCN Red List status, and residence
       (either permanent or temporary). The function will return a statement if no species were found.
}
\description{
The function takes a spatial polygon object (class "sf") representing the study area and
returns a data frame with three columns: scientific name, IUCN Red List status, and
residence (either permanent or temporary). The function downloads GBIF data for the species
present within the study area or located nearby and determines the associated IUCN status.
}
\references{
More information about GBIF can be found at https://www.gbif.org/
}
