% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Natura2000.R
\name{Plot_Natura2000}
\alias{Plot_Natura2000}
\title{Extract Natura2000 species occurrence data from a polygon and get IUCN Red List status}
\usage{
Plot_Natura2000(polygon, Natura, species, file_name)
}
\arguments{
\item{polygon}{A spatial polygon object (class "sf") representing the study area.}

\item{Natura}{An "sf" object representing the Natura2000 sites.}

\item{species}{A CSV file containing the species associated with the Natura2000 sites.}

\item{file_name}{A character string representing the name of the output file.}
}
\value{
A data frame (if species are present) with two columns: scientific name and IUCN Red List status.
}
\description{
This function takes a spatial polygon object (class "sf") representing a study area and
searches for any species that are either present within the polygon or located nearby
Natura2000 sites. It then determines the associated IUCN Red List status of the
identified species.
}
\references{
More information about Natura2000 can be found at https://ec.europa.eu/environment/nature/natura2000/
}
