% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_SpLink.R
\name{Plot_SpLink}
\alias{Plot_SpLink}
\title{Extract SpeciesLink species occurrence data from a polygon and get IUCN Red List status}
\usage{
Plot_SpLink(polygon, occurrences, file_name)
}
\arguments{
\item{polygon}{A spatial polygon object (class "sf") representing the study area.}

\item{occurrences}{A dataframe with SpeciesLink occurrence data for your region}

\item{filename}{A character string representing the name of the output file}
}
\value{
A data frame (if species are present) with two columns: scientific name and IUCN Red List status
}
\description{
This function extracts SpeciesLink species occurrence data from a polygon and retrieves
the IUCN Red List status for each species found. The output is a data frame
with columns for species name and IUCN Red List status.
}
\references{
More information about SpeciesLink can be found at https://specieslink.net/
}
